medical_back/requirements.txt
appdirs==1.4.4
APScheduler==3.6.3
asgiref==3.7.2
autopep8==2.0.4
cachetools==4.2.2
certifi==2023.11.17
charset-normalizer==3.3.2
Django==4.2.9
django-cors-headers==4.3.1
django-modeltranslation==0.18.11
django-ordered-model==3.7.4
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
fs==2.4.16
gunicorn==21.2.0
idna==3.6
importlib-metadata==7.0.1
Markdown==3.5.2
packaging==23.2
pep8==1.7.1
pillow==10.2.0
psycopg2-binary==2.9.9
pycodestyle==2.11.1
PyJWT==2.9.0
python-dotenv==1.0.0
python-telegram-bot==13.15
pytz==2023.3.post1
PyYAML==6.0.1
requests==2.31.0
sentry-sdk==1.43.0
six==1.16.0
sqlparse==0.4.4
tomli==2.0.1
tornado==6.1
typing_extensions==4.9.0
tzlocal==5.2
urllib3==1.26.15
zipp==3.17.0
